+++ web_src/js/utils/url.js
@@ -1,0 +1,3 @@
+export function pathEscapeSegments(s) {
+  return s.split('/').map(encodeURIComponent).join('/');
+}
