+++ web_src/less/_base.less
@@ -1924,10 +1924,6 @@ footer {
   display: block;
 }
 
-[v-cloak] {
-  display: none !important;
-}
-
 .repos-search {
   padding-bottom: 0 !important;
 }
