+++ web_src/less/_repository.less
@@ -222,12 +222,6 @@
       font-size: 1.2em;
     }
 
-    .choose.reference {
-      .header .icon {
-        font-size: 1.4em;
-      }
-    }
-
     .repo-path {
 
       .section,
